+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for custom cryptocurrencies defined in the database
 - Added support for the cryptocurrency _Sky_
 - Added a new market data preset (`WITHOUT_ACTIVITIES`) in the markets section of the admin control panel to filter asset profiles without activities
 
 ### Changed
 
+- Upgraded `marked` from version `17.0.1` to `17.0.2`
 - Upgraded `ngx-markdown` from version `21.0.1` to `21.1.0`
 
 ## 2.239.0 - 2026-02-15